SCB-548 reorganization springmvc it (spring-springmvc-tests and springmvc-tests) make four independent module in order to eliminate disturbance(from gracefully shutdown)
