+++ integration-tests/springmvc-tests/springmvc-tests-general-with-springboot/src/test/resources/registry.yaml
